cat-inspect-analyzer/requirements.txt
boto3 >= 1.34.0
Pillow >= 10.0.0
reportlab >= 4.0.0
python-dotenv >= 1.0.0
anthropic >= 0.25.0